requirements.txt
# MCP SDK
mcp>=1.6.0

# HTTP Client
httpx>=0.25.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.4.0
email-validator>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0
pytest-httpx>=0.25.0
pytest-asyncio>=0.21.0

# Logging
structlog>=23.1.0

# CLI Parsing
click>=8.1.0

# Async
asyncio>=3.4.3